+++ wab/create/Create_Listbox.cpp
@@ -99,7 +99,7 @@ void Listbox::Create( HWND parent )
 		GetActualWidth(),
 		GetActualHeight(), 
 		parent,
-		(HMENU)GetID(),
+		(HMENU)(uint64_t)GetID(),
 		0,
 		0
 	);
